+++ packages/sif-common-forms-ds/src/forms/virksomhet/virksomhetUtils.ts
@@ -1,5 +1,4 @@
-import { dateToISOString, ISOStringToDate, YesOrNo } from '@navikt/sif-common-formik-ds';
-import { getNumberFromStringInput } from '@navikt/sif-common-formik-ds/src/validation/validationUtils';
+import { dateToISOString, getNumberFromNumberInputValue, ISOStringToDate, YesOrNo } from '@navikt/sif-common-formik-ds';
 import { getDate4YearsAgo, guid } from '@navikt/sif-common-utils';
 import dayjs from 'dayjs';
 import { Næringstype, Virksomhet, VirksomhetFormValues } from './types';
@@ -60,8 +59,8 @@ export const mapFormValuesToVirksomhet = (
     formValues: VirksomhetFormValues,
     id: string | undefined,
 ): Partial<Virksomhet> => {
-    const næringsinntekt = getNumberFromStringInput(formValues.næringsinntekt);
-    const inntektEtterVarigEndring = getNumberFromStringInput(
+    const næringsinntekt = getNumberFromNumberInputValue(formValues.næringsinntekt);
+    const inntektEtterVarigEndring = getNumberFromNumberInputValue(
         formValues.varigEndringINæringsinntekt_inntektEtterEndring,
     );