+++ apps/omsorgsdager-aleneomsorg-dialog/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @navikt/omsorgsdager-aleneomsorg-dialog
 
+## 2.23.0
+
+### Minor Changes
+
+-   Endre i NumberInput
+
+    -   Formatere verdi i NumberInput ved blur/fokus
+    -   Kun tillate komma som desimalskille
+    -   Ta i bruk integerValue=true på inntekt i SN
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @navikt/sif-common-formik-ds@2.1.0
+    -   @navikt/sif-common-forms-ds@15.0.0
+    -   @navikt/sif-common-soknad-ds@20.0.0
+
 ## 2.22.14
 
 ### Patch Changes