+++ packages/sif-common-forms-ds/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @navikt/sif-common-forms-ds
 
+## 15.0.0
+
+### Minor Changes
+
+-   Endre i NumberInput
+
+    -   Formatere verdi i NumberInput ved blur/fokus
+    -   Kun tillate komma som desimalskille
+    -   Ta i bruk integerValue=true på inntekt i SN
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @navikt/sif-common-formik-ds@2.1.0
+
 ## 14.0.1
 
 ### Patch Changes